+++ ktrain/text/preprocessor.py
@@ -38,6 +38,24 @@ def fname_from_url(url):
     return os.path.split(url)[-1]
 
 
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
 #------------------------------------------------------------------------------
 # Word Vectors
 #------------------------------------------------------------------------------
@@ -356,7 +374,7 @@ class TextPreprocessor(Preprocessor):
         self.lang = lang
         self.multilabel = multilabel
 
-
+    
     def get_preprocessor(self):
         raise NotImplementedError
 
@@ -370,7 +388,7 @@ class TextPreprocessor(Preprocessor):
 
 
     def set_multilabel(self, data, mode):
-        if mode == 'train':
+        if mode == 'train' and self.get_classes():
             self.multilabel = U.is_multilabel(data)
 
 
@@ -400,6 +418,37 @@ class TextPreprocessor(Preprocessor):
             seg_list = list(seg_list)
             split_texts.append(seg_list)
         return [" ".join(tokens) for tokens in split_texts] 
+
+
+    @classmethod
+    def seqlen_stats(cls, list_of_texts):
+        """
+        compute sequence length stats from
+        list of texts in any spaces-segmented language
+        Args:
+            list_of_texts: list of strings
+        Returns:
+            dict: dictionary with keys: mean, 95percentile, 99percentile
+        """
+        counts = []
+        for text in list_of_texts:
+            counts.append(len(text.split()))
+        p95 = np.percentile(counts, 95)
+        p99 = np.percentile(counts, 99)
+        avg = sum(counts)/len(counts)
+        return {'mean':avg, '95percentile': p95, '99percentile':p99}
+
+
+    def print_seqlen_stats(self, texts, mode, verbose=1):
+        """
+        prints stats about sequence lengths
+        """
+        if verbose and not self.is_nospace_lang():
+            stat_dict = TextPreprocessor.seqlen_stats(texts)
+            print( "%s sequence lengths:" % mode)
+            for k in stat_dict:
+                print("\t%s : %s" % (k, int(round(stat_dict[k]))))
+
 
 
 
@@ -454,7 +503,13 @@ class StandardTextPreprocessor(TextPreprocessor):
         # convert to word IDs
         x_train = self.tok.texts_to_sequences(train_text)
         U.vprint('{} train sequences'.format(len(x_train)), verbose=verbose)
-        U.vprint('Average train sequence length: {}'.format(np.mean(list(map(len, x_train)), dtype=int)), verbose=verbose)
+        sequence_lengths = list(map(len, x_train))
+        U.vprint('Average train sequence length: {}'.format(np.mean(sequence_lengths, dtype=int)), 
+                verbose=verbose)
+        U.vprint('95th percentile train sequence length: {}'.format(np.percentile(sequence_lengths, 95)), 
+                verbose=verbose)
+        U.vprint('99th percentile train sequence length: {}'.format(np.percentile(sequence_lengths, 99)), 
+                verbose=verbose)
 
         # add ngrams
         x_train = self._fit_ngrams(x_train, verbose=verbose)
@@ -487,8 +542,13 @@ class StandardTextPreprocessor(TextPreprocessor):
         # convert to word IDs
         x_test = self.tok.texts_to_sequences(test_text)
         U.vprint('{} test sequences'.format(len(x_test)), verbose=verbose)
-        U.vprint('Average test sequence length: {}'.format(np.mean(list(map(len, x_test)), 
+        sequence_lengths = list(map(len, x_test))
+        U.vprint('Average test sequence length: {}'.format(np.mean(sequence_lengths, 
                                                                  dtype=int)), verbose=verbose)
+        U.vprint('95th percentile test sequence length: {}'.format(np.percentile(sequence_lengths, 95)), 
+                  verbose=verbose)
+        U.vprint('99th percentile test sequence length: {}'.format(np.percentile(sequence_lengths, 99)), 
+                  verbose=verbose)
 
         # add n-grams
         x_test = self._add_ngrams(x_test, mode='test', verbose=verbose)
@@ -701,12 +761,14 @@ class TransformersPreprocessor(TextPreprocessor):
         """
         U.vprint('preprocessing %s...' % (mode), verbose=verbose)
         U.vprint('language: %s' % (self.lang), verbose=verbose)
+        self.print_seqlen_stats(texts, mode, verbose=verbose)
+
         if y is None and mode == 'train':
             raise ValueError('y is required for training sets')
         elif y is None:
             y = np.array([1] * len(texts))
         if y is not None:
-            if len(y.shape) == 1:
+            if len(y.shape) == 1 and self.get_classes():
                 y = to_categorical(y)
 
         dataset = hf_convert_examples(texts, y=y, tokenizer=self.tok, max_length=self.maxlen,
@@ -775,8 +837,8 @@ class Transformer(TransformersPreprocessor):
         """
         if not use_with_learner and batch_size is None:
             raise ValueError('batch_size is required when use_with_learner=False')
-        if classes is None or not classes:
-            raise ValueError('classes argument is required - provide list of class names as strings')
+        if multilabel and (classes is None or not classes):
+            raise ValueError('classes argument is required when multilabel=True')
         super().__init__(model_name,
                          maxlen, max_features=10000, classes=classes, multilabel=multilabel)
         self.batch_size = batch_size
@@ -827,6 +889,9 @@ class Transformer(TransformersPreprocessor):
 
 
     def get_classifier(self):
+        if not self.get_classes():
+            warnings.warn('no class labels were provided - treating as regression')
+            return self.get_regressor()
         num_labels = len(self.get_classes())
         model = self.model_type.from_pretrained(self.model_name, num_labels=num_labels)
         if self.multilabel:
@@ -837,6 +902,25 @@ class Transformer(TransformersPreprocessor):
                       optimizer=keras.optimizers.Adam(learning_rate=3e-5, epsilon=1e-08),
                       metrics=['accuracy'])
         return model
+
+
+    def get_regressor(self):
+        if self.get_classes():
+            warnings.warn('class labels were provided - treating as classification problem')
+            return self.get_classifier()
+        model = self.model_type.from_pretrained(self.model_name, num_labels=1)
+        loss_fn = 'mse'
+        model.compile(loss=loss_fn,
+                      optimizer=keras.optimizers.Adam(learning_rate=3e-5, epsilon=1e-08),
+                      metrics=['mae'])
+        return model
+
+
+    def get_model(self):
+        if not self.get_classes():
+            return self.get_regressor()
+        else:
+            return self.get_classifier()